[react] show ingredients using hooks
