+++ burger-stacker/src/Ingredients/IngredientsList.js
@@ -6,16 +6,29 @@ let IngredientsList = (props) => {
 
     const[burger, setBurger] = useState([])
 
+    const ingredients = [
+        {name: 'Kaiser Bun', color: 'saddlebrown'},
+        {name: 'Sesame Bun', color: 'sandybrown'},
+        {name: 'Gluten Free Bun', color: 'peru'},
+        {name: 'Lettuce Wrap', color: 'olivedrab'},
+        {name: 'Beef Patty', color: '#3F250B'},
+        {name: 'Soy Patty', color: '#3F250B'},
+        {name: 'Black Bean Patty', color: '#3F250B'},
+        {name: 'Chicken Patty', color: 'burlywood'},
+        {name: 'Lettuce', color: 'lawngreen'},
+        {name: 'Tomato', color: 'tomato'},
+        {name: 'Bacon', color: 'maroon'},
+        {name: 'Onion', color: 'lightyellow'}
+      ]
 
-    const addIngredients = (e, ingredients) => {
-        e.preventDefault()
-        setBurger(burger.concat(ingredient))
+    const addIngredients = (ingredients) => {
+        setBurger(burger.concat(ingredients))
     }
 
         return (
             <div>
                 <h1>Ingredients</h1>
-                {/* <ul>{ingredientsList}</ul> */}
+                <ul><Ingredients ingredients={ingredients} addIngredients={addIngredients}/></ul>
                 <h1>Burger</h1>
                 {/* <ul>{burgerList}</ul> */}
             </div>
