+++ burger-stacker/src/index.js
@@ -5,25 +5,11 @@ import Burger from './Burger/BurgerPane';
 import IngredientsList from './Ingredients/IngredientsList';
 import BurgerStack from './Burger/BurgerStack'
 
-const ingredients = [
-  {name: 'Kaiser Bun', color: 'saddlebrown'},
-  {name: 'Sesame Bun', color: 'sandybrown'},
-  {name: 'Gluten Free Bun', color: 'peru'},
-  {name: 'Lettuce Wrap', color: 'olivedrab'},
-  {name: 'Beef Patty', color: '#3F250B'},
-  {name: 'Soy Patty', color: '#3F250B'},
-  {name: 'Black Bean Patty', color: '#3F250B'},
-  {name: 'Chicken Patty', color: 'burlywood'},
-  {name: 'Lettuce', color: 'lawngreen'},
-  {name: 'Tomato', color: 'tomato'},
-  {name: 'Bacon', color: 'maroon'},
-  {name: 'Onion', color: 'lightyellow'}
-]
+
 
 ReactDOM.render(
   <React.StrictMode>
-    <Burger />
-    <IngredientsList ingredients={ingredients}/>
+    <IngredientsList />
   </React.StrictMode>,
   document.getElementById('root')
 );
